Conflicts:
	include/llvm/ADT/Triple.h
	lib/Bitcode/NaCl/Analysis/NaClBitcodeValueDist.cpp
	lib/Support/Triple.cpp
	lib/Target/CppBackend/CPPTargetMachine.h
	lib/Transforms/IPO/PassManagerBuilder.cpp
	lib/Transforms/NaCl/ExpandArithWithOverflow.cpp
	lib/Transforms/NaCl/ExpandVarArgs.cpp
	lib/Transforms/NaCl/FlattenGlobals.cpp
	lib/Transforms/NaCl/PNaClABISimplify.cpp
	lib/Transforms/NaCl/PromoteIntegers.cpp
	lib/Transforms/NaCl/ResolvePNaClIntrinsics.cpp
	test/Transforms/NaCl/promote-integers.ll
	tools/CMakeLists.txt
	tools/LLVMBuild.txt
	tools/bugpoint/LLVMBuild.txt
	tools/bugpoint/Makefile
	tools/llvm-vtabledump/Makefile
	tools/opt/CMakeLists.txt
	unittests/Bitcode/Makefile

+++ unittests/Bitcode/NaClAbbrevTrieTest.cpp
@@ -131,13 +131,13 @@ TEST(NaClAbbrevTrieTest, Simple) {
   // Build lookup map, and check that we build the expected trie.
   AbbrevLookupSizeMap LookupMap;
   NaClBuildAbbrevLookupMap(LookupMap, Abbrevs);
-  EXPECT_EQ(1, LookupMap.size())
+  EXPECT_EQ((size_t)1, LookupMap.size())
       << "There should only be one entry in the Lookup map "
       << "for abbreviations of length 2";
   for (AbbrevLookupSizeMap::iterator
            Iter = LookupMap.begin(), IterEnd = LookupMap.end();
        Iter != IterEnd; ++Iter) {
-    EXPECT_EQ(Iter->first, 2)
+    EXPECT_EQ(Iter->first, (size_t)2)
         << "Expecting abbreviations to be of length 2";
     EXPECT_EQ(std::string(
         "Abbreviations:\n"
@@ -350,7 +350,7 @@ TEST(NaClAbbrevTrieTest, Array) {
   for (AbbrevLookupSizeMap::iterator
            Iter = LookupMap.begin(), IterEnd = LookupMap.end();
        Iter != IterEnd; ++Iter) {
-    EXPECT_LE(0, Iter->first);
+    EXPECT_LE((size_t)0, Iter->first);
     EXPECT_GE(MaxValueIndex, Iter->first);
     if (Iter->first == 5) {
       // Note that all abbreviations accept records with 5 values.
